Conflicts:
	Common/version.txt
	b3270/b3270.man
	b3270/x3270-script.man
	c3270/c3270.man
	c3270/ibm_hosts.man
	c3270/x3270-script.man
	mitm/mitm.man
	pr3287/pr3287.man
	s3270/s3270.man
	s3270/x3270-script.man
	tcl3270/tcl3270.man
	x3270/ibm_hosts.man
	x3270/x3270-script.man
	x3270/x3270.man
	x3270if/x3270if.man

+++ wc3270/relink.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2006-2009, 2013-2015, 2019 Paul Mattes.
+ * Copyright (c) 2006-2009, 2013-2015, 2019-2020 Paul Mattes.
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without
@@ -322,7 +322,7 @@ read_session(FILE *f, session_t *s, char **usp)
 		}
 		if (s_off > sizeof(*s)) {
 #if defined(DEBUG_EDIT) /*[*/
-		    printf("[s overflow: %d > %d]\n", s_off, sizeof(*s));
+		    printf("[s overflow: %d > %d]\n", s_off, (int)sizeof(*s));
 #endif /*]*/
 		    return 0;
 		}
